+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/DefaultAudioVideoFacade.kt
@@ -8,6 +8,8 @@ package com.amazonaws.services.chime.sdk.meetings.audiovideo
 import android.Manifest
 import android.content.Context
 import android.content.pm.PackageManager
+import android.os.Build
+import androidx.annotation.RequiresApi
 import androidx.core.content.ContextCompat
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.activespeakerdetector.ActiveSpeakerDetectorFacade
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.activespeakerdetector.ActiveSpeakerObserver
@@ -17,8 +19,8 @@ import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.VideoRenderVie
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.VideoSource
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.VideoTileController
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.VideoTileObserver
+import com.amazonaws.services.chime.sdk.meetings.device.DefaultDeviceController
 import com.amazonaws.services.chime.sdk.meetings.device.DeviceChangeObserver
-import com.amazonaws.services.chime.sdk.meetings.device.DeviceController
 import com.amazonaws.services.chime.sdk.meetings.device.MediaDevice
 import com.amazonaws.services.chime.sdk.meetings.realtime.RealtimeControllerFacade
 import com.amazonaws.services.chime.sdk.meetings.realtime.RealtimeObserver
@@ -28,7 +30,7 @@ class DefaultAudioVideoFacade(
     private val context: Context,
     private val audioVideoController: AudioVideoControllerFacade,
     private val realtimeController: RealtimeControllerFacade,
-    private val deviceController: DeviceController,
+    private val deviceController: DefaultDeviceController,
     private val videoTileController: VideoTileController,
     private val activeSpeakerDetector: ActiveSpeakerDetectorFacade
 ) : AudioVideoFacade {
@@ -74,6 +76,7 @@ class DefaultAudioVideoFacade(
 
     override fun stop() {
         audioVideoController.stop()
+        deviceController.stopListening()
     }
 
     override fun startLocalVideo() {
@@ -138,6 +141,11 @@ class DefaultAudioVideoFacade(
 
     override fun chooseAudioDevice(mediaDevice: MediaDevice) {
         deviceController.chooseAudioDevice(mediaDevice)
+    }
+
+    @RequiresApi(Build.VERSION_CODES.N)
+    override fun getActiveAudioDevice(): MediaDevice? {
+        return deviceController.getActiveAudioDevice()
     }
 
     override fun getActiveCamera(): MediaDevice? {