+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/device/DeviceChangeObserver.kt
@@ -19,4 +19,13 @@ interface DeviceChangeObserver {
      * @param freshAudioDeviceList: List<[MediaDevice]> - An updated list of audio devices.
      */
     fun onAudioDeviceChanged(freshAudioDeviceList: List<MediaDevice>)
+
+    /**
+     * Called when chooseAudioDevice is called.
+     *
+     * Note: this callback will be called on main thread.
+     *
+     * @param device: [MediaDevice] - A device [DeviceController.chooseAudioDevice] is called with
+     */
+    fun onChooseAudioDeviceCalled(device: MediaDevice) {}
 }