+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/device/DeviceController.kt
@@ -5,6 +5,9 @@
 
 package com.amazonaws.services.chime.sdk.meetings.device
 
+import android.os.Build
+import androidx.annotation.RequiresApi
+
 /**
  * [DeviceController] keeps track of the devices being used for audio device
  * (e.g. built-in speaker), video input (e.g. camera)).
@@ -27,6 +30,15 @@ interface DeviceController {
      */
     fun chooseAudioDevice(mediaDevice: MediaDevice)
 
+    /**
+     * Get the active local audio device in the meeting, return null if there isn't any.
+     *
+     * NOTE: This requires Android API 24 and above
+     *
+     * @return the active local audio device
+     */
+    @RequiresApi(Build.VERSION_CODES.N)
+    fun getActiveAudioDevice(): MediaDevice?
     /**
      * Adds an observer to receive callbacks about device changes.
      *