+++ CHANGELOG.md
@@ -52,6 +52,12 @@ This release includes support for custom video sources, and therefore includes a
 ### Changed
 * Changed `MAX_TILE_COUNT` in the demo app from 4 to 16. Now the demo app can support at most 16 remote video tiles.
 
+### Added
+* Added `getActiveAudioDevice` in `DefaultDeviceController`
+
+### Fixed
+* Fixed `DefaultDeviceController` returning A2DP bluetooth devices
+
 ## [0.7.3] - 2020-09-10
 
 ### Fixed